# Server/swagger.yaml
swagger: "2.0"
info:
  description: "API của Website Hỗ trợ đào tạo SGU"
  version: "1.0.0"
  title: "Website Hỗ trợ đào tạo SGU"
  contact:
    email: ""
host: ""
...
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /admin/tai-khoan/DanhSachTK:
    get: 
      tags:
        - "TaiKhoan"
      summary: Admin lấy danh sách tài khoản
      produces:
        - application/json     
      parameters:
      - in: query
        name: page
        type: integer
        required: true
        description: số trang (ví dụ trang 1, trang 2), giá trị > 0
      - in: query
        name: pageSize
        type: integer
        required: true
        description: giới hạn số lượng 1 trang, giá trị > 0
      - in: query
        name: keyword
        type: string
        required: true
        description: mã tài khoản hoặc tên đăng nhập hoặc mã quyền tài khoản
      responses:
        200: 
          description: Lấy danh sách tài khoản thành công
          schema:
            type: object
            example: {
              "status": true,
              "message": "Lấy danh sách tài khoản thành công",
            }
        400:
          description: Lấy danh sách tài khoản thất bại
          schema:
            type: object
            example: {
              "status": false,
              "message": "Lấy danh sách tài khoản thất bại",
            }
      security:
      - ApiKeyAuth: []
  /admin/tai-khoan/ChiTietTaiKhoan/{MaTK}:
    get:
      tags:
        - "TaiKhoan"
      summary: "Chi tiết thông tin tài khoản"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
      responses:
        200:
          description: Chi tiết thông tin tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chi tiết thông tin tài khoản thành công.",
            }
        400:
          description: Chi tiết thông tin tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chi tiết thông tin tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/tai-khoan/Them:
    post: 
      tags:
        - "TaiKhoan"
      summary: Thêm tài khoản
      produces:
        - application/json     
      parameters:
        - in: body
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
        - in: body
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
        - in: body
          name: TenDangNhap
          type: string
          required: true
          description: Tên đăng nhập
        - in: body
          name: MatKhau
          type: string
          description: Mật khẩu đăng nhập
      responses:
        200:
          description: Thêm tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Thêm tài khoản thành công.",
            }          
        400:
          description: Thêm tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Thêm tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: [] 
  /admin/tai-khoan/ChinhSua/{MaTK}:
    put: 
      tags:
        - "TaiKhoan"
      summary: Chỉnh sửa tài khoản
      produces:
        - application/json     
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
        - in: body
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
        - in: body
          name: TrangThai
          type: string
          required: true
          enum: ['Chưa kích hoạt','Đã kích hoạt']
          description: Trạng thái tài khoản
      responses:
        200:
          description: Chỉnh sửa tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chỉnh sửa tài khoản thành công.",
            }          
        400:
          description: Chỉnh sửa tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Chỉnh sửa tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []     
  /admin/tai-khoan/Xoa/{MaTK}:
    delete:
      tags:
        - "TaiKhoan"
      summary: "Xóa thông tin tài khoản"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
      responses:
        200:
          description: Xóa thông tin tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Xóa thông tin tài khoản thành công.",
            }
        400:
          description: Xóa thông tin tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Xóa thông tin tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /admin/tai-khoan/DanhSachTKGiangVien:
    get: 
      tags:
        - "TaiKhoan"
      summary: Admin lấy danh sách tài khoản giảng viên
      produces:
        - application/json     
      parameters:
      - in: query
        name: page
        type: integer
        description: số trang (ví dụ trang 1, trang 2), giá trị > 0
      - in: query
        name: pageSize
        type: integer
        description: giới hạn số lượng 1 trang, giá trị > 0
      responses:
        200: 
          description: Lấy danh sách tài khoản giảng viên thành công
          schema:
            type: object
            example: {
              "status": true,
              "message": "Lấy danh sách tài khoản giảng viên thành công",
            }
        400:
          description: Lấy danh sách tài khoản giảng viên thất bại
          schema:
            type: object
            example: {
              "status": false,
              "message": "Lấy danh sách tài khoản giảng viên thất bại",
            }
      security:
      - ApiKeyAuth: []
  /admin/tai-khoan/DanhSachTKSinhVien:
    get: 
      tags:
        - "TaiKhoan"
      summary: Admin lấy danh sách tài khoản sinh viên
      produces:
        - application/json     
      parameters:
      - in: query
        name: page
        type: integer
        description: số trang (ví dụ trang 1, trang 2), giá trị > 0
      - in: query
        name: pageSize
        type: integer
        description: giới hạn số lượng 1 trang, giá trị > 0
      responses:
        200: 
          description: Lấy danh sách tài khoản sinh viên thành công
          schema:
            type: object
            example: {
              "status": true,
              "message": "Lấy danh sách tài khoản sinh viên thành công",
            }
        400:
          description: Lấy danh sách tài khoản sinh viên thất bại
          schema:
            type: object
            example: {
              "status": false,
              "message": "Lấy danh sách tài khoản sinh viên thất bại",
            }
      security:
      - ApiKeyAuth: []
...
  /admin/tai-khoan/PhucHoiMatKhau/{MaTK}:
    post: 
      tags:
        - "TaiKhoan"
      summary: Phục hồi mật khẩu
      produces:
        - application/json     
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
        - in: body
          name: MatKhauPhucHoi
          type: string
          required: true
          description: Mật khẩu phục hồi
      responses:
        200:
          description: Phục hồi mật khẩu thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Phục hồi mật khẩu thành công.",
            }          
        400:
          description: Phục hồi mật khẩu thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Phục hồi mật khẩu thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /admin/tai-khoan/KichHoatTaiKhoan/{MaTK}:
    post: 
      tags:
        - "TaiKhoan"
      summary: Kích hoạt tài khoản
      produces:
        - application/json     
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
      responses:
        200:
          description: Kích hoạt tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Kích hoạt tài khoản thành công.",
            }          
        400:
          description: Kích hoạt tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Kích hoạt tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []  
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: headers
    name: x-access-token